fix slow loading weight, vaccine, service edit popups
